NExt/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
transformers
torch
sentence-transformers
python-multipart
soundfile
